Cleanup code for up-coming 2.2 release, fixing warnings etc on windows.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@662527 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/marshal/PrimitiveMapMarshaller.cpp
@@ -141,7 +141,7 @@ void PrimitiveMapMarshaller::marshalPrimitiveList(
         throw ( decaf::io::IOException ) {
 
     try{
-        dataOut.writeInt( list.size() );
+        dataOut.writeInt( (int)list.size() );
 
         for( std::size_t ix = 0; ix < list.size(); ++ix ) {
             marshalPrimitive( dataOut, list.get( ix ) );
